# Use Fused Multiply Add instead of normal arithmetic operations