a * b + c operations can be done in one operation. It's even standard C99.
Let's benchmark this.