@thedrow 2016-09-01
# Original
```
-------------------------------------------------------------------------------------------------------------------------------------- benchmark 'decode': 49 tests --------------------------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                                                                                     Min                     Max               Mean                StdDev             Median               IQR            Outliers(*)  Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_decode_cffi[jR-4]                                                                                                                             4.0770 (1.0)       47,486.8059 (>1000.0)   7.1814 (1.49)       305.8285 (186.48)    4.3869 (1.0)      0.3099 (inf)           5;2385   24529          10
test_decode_cffi[ejRe-4]                                                                                                                           4.1008 (1.01)         100.0881 (2.97)      4.8350 (1.0)          2.0014 (1.22)      4.3869 (1.0)      0.2146 (inf)        1352;3355   24386          10
test_decode_cffi[olejRejN-4]                                                                                                                       4.1962 (1.03)          39.2914 (1.17)      4.9456 (1.02)         1.8413 (1.12)      4.4107 (1.01)     0.2146 (inf)        1387;3878   23832          10
test_decode_cffi[2VolejRejNmG-4]                                                                                                                   4.2915 (1.05)          33.6885 (1.0)       4.9752 (1.03)         1.7390 (1.06)      4.5061 (1.03)     0.1192 (inf)        1338;4584   23302          10
test_decode_cffi[ejRe-0]                                                                                                                           4.2915 (1.05)          35.2144 (1.05)      5.0863 (1.05)         2.0584 (1.26)      4.5061 (1.03)     0.2146 (inf)        1506;3821   23832          10
test_decode_cffi[jR-0]                                                                                                                             4.2915 (1.05)          89.5977 (2.66)      5.0881 (1.05)         2.0783 (1.27)      4.5061 (1.03)     0.4053 (inf)        1683;2024   23832          10
test_decode_cffi[olejRejN-0]                                                                                                                       4.3869 (1.08)          43.4160 (1.29)      5.0138 (1.04)         1.8065 (1.10)      4.5776 (1.04)     0.0954 (inf)        1370;4513   22796          10
test_decode_cffi[2VolejRejNmG-0]                                                                                                                   4.5776 (1.12)         175.8099 (5.22)      5.3844 (1.11)         2.2917 (1.40)      4.7922 (1.09)     0.2861 (inf)        1157;3730   22193          10
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-4]                                                                                                       4.6015 (1.13)          50.7832 (1.51)      5.2681 (1.09)         1.6400 (1.0)       4.8161 (1.10)     0.1192 (inf)        1369;4362   21846          10
test_decode_cffi[2VolejRejNmG-12]                                                                                                                  4.8876 (1.20)      13,849.9737 (411.12)    7.3382 (1.52)        55.2928 (33.71)     5.4836 (1.25)     0.4768 (inf)        369;17894  102301           2
test_decode_cffi[2VolejRejNmG-24]                                                                                                                  4.8876 (1.20)       1,176.9533 (34.94)     6.2555 (1.29)         8.0544 (4.91)      5.4836 (1.25)     0.1192 (inf)       1223;32957  102301           2
test_decode_cffi[2VolejRejNmG-48]                                                                                                                  4.8876 (1.20)     543,997.5262 (>1000.0)  12.2422 (2.53)     1,715.1642 (>1000.0)   5.4836 (1.25)     0.4768 (inf)           3;4154  102301           2
test_decode_cffi[2VolejRejNmG-8]                                                                                                                   4.8876 (1.20)     548,679.4710 (>1000.0)  12.3611 (2.56)     1,728.4333 (>1000.0)   5.4836 (1.25)     0.1192 (inf)          4;41969  102301           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-4]                                                                               4.8876 (1.20)         188.4699 (5.59)      6.0475 (1.25)         3.5985 (2.19)      5.4836 (1.25)     0.1192 (inf)       3019;29235   99865           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-0]                                                                                                       4.8876 (1.20)         685.5726 (20.35)     5.9689 (1.23)         4.0199 (2.45)      5.4836 (1.25)     0.1192 (inf)       2291;28872  102301           2
test_decode_cffi[ejRe-12]                                                                                                                          4.8876 (1.20)         231.5044 (6.87)      5.8821 (1.22)         3.9705 (2.42)      5.4836 (1.25)     0.4768 (inf)        3946;6949  102301           2
test_decode_cffi[ejRe-24]                                                                                                                          4.8876 (1.20)     415,034.4133 (>1000.0)  10.5719 (2.19)     1,309.9398 (798.73)    5.4836 (1.25)     0.4768 (inf)           5;6944  102301           2
test_decode_cffi[ejRe-48]                                                                                                                          4.8876 (1.20)     407,080.4119 (>1000.0)  10.6045 (2.19)     1,281.9739 (781.68)    5.4836 (1.25)     0.1192 (inf)          6;39305  102301           2
test_decode_cffi[ejRe-8]                                                                                                                           4.8876 (1.20)         219.4643 (6.51)      5.8106 (1.20)         4.0597 (2.48)      5.4836 (1.25)     0.4768 (inf)        3099;5811  102301           2
test_decode_cffi[jR-12]                                                                                                                            4.8876 (1.20)         507.9508 (15.08)     5.9239 (1.23)         4.9237 (3.00)      5.4836 (1.25)     0.4768 (inf)        1905;7821  102301           2
test_decode_cffi[jR-24]                                                                                                                            4.8876 (1.20)         204.4439 (6.07)      5.7390 (1.19)         3.4888 (2.13)      5.4836 (1.25)     0.4768 (inf)        3419;5131  102301           2
test_decode_cffi[jR-48]                                                                                                                            4.8876 (1.20)         144.0048 (4.27)      5.8224 (1.20)         3.4754 (2.12)      5.4836 (1.25)     0.4768 (inf)        2579;5861  102301           2
test_decode_cffi[jR-8]                                                                                                                             4.8876 (1.20)         282.5260 (8.39)      5.8201 (1.20)         4.0692 (2.48)      5.4836 (1.25)     0.4768 (inf)        3044;6257  102301           2
test_decode_cffi[olejRejN-12]                                                                                                                      4.8876 (1.20)         157.5947 (4.68)      5.8357 (1.21)         3.7236 (2.27)      5.4836 (1.25)     0.4768 (inf)        2400;6817  102301           2
test_decode_cffi[olejRejN-24]                                                                                                                      4.8876 (1.20)       1,993.5369 (59.18)     5.9199 (1.22)         7.4194 (4.52)      5.4836 (1.25)     0.0000 (1.0)       1012;48042  102301           2
test_decode_cffi[olejRejN-48]                                                                                                                      4.8876 (1.20)         665.0686 (19.74)     6.0245 (1.25)         4.2170 (2.57)      5.4836 (1.25)     0.1192 (inf)       3381;32520  102301           2
test_decode_cffi[olejRejN-8]                                                                                                                       4.8876 (1.20)     501,541.9722 (>1000.0)  11.5149 (2.38)     1,580.3009 (963.58)    5.4836 (1.25)     0.4768 (inf)           4;8172  102301           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-12]                                                                                                      5.4836 (1.35)         159.0252 (4.72)      6.2508 (1.29)         3.3717 (2.06)      5.9605 (1.36)     0.5960 (inf)        2725;5449   91181           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-24]                                                                                                      5.4836 (1.35)         325.9182 (9.67)      6.3871 (1.32)         3.8980 (2.38)      5.9605 (1.36)     0.5960 (inf)        2606;6184   91181           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-48]                                                                                                      5.4836 (1.35)         128.0308 (3.80)      6.5572 (1.36)         3.7681 (2.30)      5.9605 (1.36)     0.1192 (inf)       3553;36680   91181           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-8]                                                                                                       5.4836 (1.35)         351.5482 (10.44)     6.2442 (1.29)         3.6115 (2.20)      5.9605 (1.36)     0.5960 (inf)        2661;4308   91181           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-0]                                                                               5.9605 (1.46)         168.5619 (5.00)      7.3225 (1.51)         3.8065 (2.32)      6.5565 (1.49)     0.5960 (inf)        3034;6947   83887           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-12]                                                                              5.9605 (1.46)         152.5879 (4.53)      7.1140 (1.47)         3.8662 (2.36)      6.5565 (1.49)     0.8345 (inf)        3861;5677   83887           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-24]                                                                              5.9605 (1.46)         233.5310 (6.93)      6.9715 (1.44)         3.7741 (2.30)      6.4373 (1.47)     0.4768 (inf)        2592;8024   83887           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-48]                                                                              5.9605 (1.46)         122.9048 (3.65)      7.1111 (1.47)         3.5391 (2.16)      6.5565 (1.49)     0.4768 (inf)        2836;4686   83887           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-8]                                                                               5.9605 (1.46)       1,783.0133 (52.93)     7.0176 (1.45)         7.6655 (4.67)      6.4373 (1.47)     0.4768 (inf)        1303;7591   83887           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-4]       6.4373 (1.58)         151.5150 (4.50)      7.8218 (1.62)         3.5034 (2.14)      7.0333 (1.60)     0.4768 (inf)        3119;6225   77673           2
test_decode_cffi[2VolejRejNmG-120]                                                                                                                 6.9141 (1.70)          97.5132 (2.89)      8.2381 (1.70)         3.6192 (2.21)      7.5102 (1.71)     0.4768 (inf)        3191;5526   72316           2
test_decode_cffi[ejRe-120]                                                                                                                         6.9141 (1.70)       1,906.9910 (56.61)     9.2623 (1.92)        17.6516 (10.76)     7.5102 (1.71)     2.5034 (inf)         653;1636   72316           2
test_decode_cffi[jR-120]                                                                                                                           6.9141 (1.70)         244.0214 (7.24)      7.8417 (1.62)         3.6782 (2.24)      7.5102 (1.71)     0.4768 (inf)        1892;6088   72316           2
test_decode_cffi[olejRejN-120]                                                                                                                     6.9141 (1.70)         188.9467 (5.61)      8.1940 (1.69)         4.2140 (2.57)      7.5102 (1.71)     0.0000 (1.0)       2782;32550   72316           2
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-120]                                                                                                     7.3910 (1.81)         154.4952 (4.59)      8.2822 (1.71)         3.3540 (2.05)      7.9870 (1.82)     0.4768 (inf)        1528;4350   67651           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-12]      7.3910 (1.81)         451.5648 (13.40)     8.9504 (1.85)         4.4236 (2.70)      7.9870 (1.82)     0.4768 (inf)        2898;8493   67651           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-24]      7.3910 (1.81)         182.5094 (5.42)      8.9270 (1.85)         3.9565 (2.41)      7.9870 (1.82)     0.4768 (inf)        3828;8125   67651           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-48]      7.3910 (1.81)         120.9974 (3.59)      8.8373 (1.83)         3.5530 (2.17)      7.9870 (1.82)     0.4768 (inf)        2876;6375   67651           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-8]       7.3910 (1.81)     566,915.5121 (>1000.0)  18.6661 (3.86)     2,201.9193 (>1000.0)   7.9870 (1.82)     0.4768 (inf)           4;6024   67651           2
test_decode_cffi[KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwE-120]                                                                             7.9870 (1.96)      11,693.4776 (347.11)   10.3046 (2.13)        72.3513 (44.12)     8.4639 (1.93)     0.4768 (inf)         100;7506   62602           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-0]       8.9407 (2.19)         128.9845 (3.83)     10.4120 (2.15)         4.1208 (2.51)      9.5367 (2.17)     0.4768 (inf)        2564;5930   55925           2
test_decode_cffi[W9jOVvKy2G7qgRLGgV6QYRoBlrkyXZ5A43v2KAyZnxO60XzkL39J4q2VolejRejNmGQBW71gPv58RYDMrpwEO8z9jD0MEmNJqK7nw1xWPpBxkm5A4nYzXoMZ-120]     9.4175 (2.31)          93.5793 (2.78)     10.9536 (2.27)         3.7253 (2.27)     10.0136 (2.28)     0.4768 (inf)        2196;5578   50534           2
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

# This PR
```
--------------------------------------------------------------------------------------------------------------------------------------- benchmark 'decode': 49 tests --------------------------------------------------------------------------------------------------------------------------------------
Name (time in us)                                                                                                                                      Min                     Max               Mean                StdDev             Median               IQR            Outliers(*)  Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_decode_cffi[ejRe-4]                                                                                                                            4.1962 (1.0)           52.6905 (1.34)      4.8391 (1.0)          1.7061 (1.0)       4.4107 (1.0)      0.1192 (inf)        1268;4321   23832          10
test_decode_cffi[jR-4]                                                                                                                              4.1962 (1.0)       74,549.4127 (>1000.0)   9.1162 (1.88)       486.6716 (285.25)    4.4107 (1.0)      0.6199 (inf)           5;4708   23832          10
test_decode_cffi[jR-0]                                                                                                                              4.2915 (1.02)         512.5999 (13.08)     5.2181 (1.08)         4.5546 (2.67)      4.5061 (1.02)     0.1907 (inf)         901;3647   23832          10
test_decode_cffi[olejRejN-4]                                                                                                                        4.2915 (1.02)          56.4098 (1.44)      5.0315 (1.04)         1.8908 (1.11)      4.5061 (1.02)     0.1907 (inf)        1449;3627   23302          10
test_decode_cffi[2VolejRejNmG-4]                                                                                                                    4.3869 (1.05)       4,093.8854 (104.45)    6.2022 (1.28)        36.5917 (21.45)     4.6015 (1.04)     0.2861 (inf)          67;3190   22796          10
test_decode_cffi[ejRe-0]                                                                                                                            4.3869 (1.05)          54.4071 (1.39)      5.0531 (1.04)         1.7422 (1.02)      4.6015 (1.04)     0.0954 (inf)        1221;4135   22796          10
test_decode_cffi[olejRejN-0]                                                                                                                        4.3869 (1.05)          41.3179 (1.05)      5.1371 (1.06)         1.7855 (1.05)      4.6015 (1.04)     0.2861 (inf)        1401;2396   22796          10
test_decode_cffi[2VolejRejNmG-0]                                                                                                                    4.6968 (1.12)         191.7839 (4.89)      6.3122 (1.30)         2.8151 (1.65)      5.1975 (1.18)     2.1935 (inf)         1284;807   21291          10
test_decode_cffi[L39J4q2VolejRejNmGQBW71g-4]                                                                                                        4.6968 (1.12)          39.1960 (1.0)       5.4360 (1.12)         1.7730 (1.04)      4.9114 (1.11)     0.3099 (inf)        1496;2174   21291          10
test_decode_cffi[2VolejRejNmG-12]                                                                                                                   4.8876 (1.16)         163.0783 (4.16)      6.0896 (1.26)         3.5759 (2.10)      5.4836 (1.24)     0.4768 (inf)        3385;5395  102301           2
…